requirements.txt
tqdm==4.66.1
Pillow==10.0.1
numpy==1.26.0
pandas==2.1.3
torch==2.1.0
lightning==2.1.2
torchmetrics==1.2.0
torchvision==0.16.0
PyYAML==6.0
milwsi==0.0.1
setuptools==68.0.0
scikit-learn==1.3.2